# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robo Arena API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RobotPartCategory =
  (typeof RobotPartCategory)[keyof typeof RobotPartCategory];

export const RobotPartCategory = {
  body: "body",
  attack: "attack",
  defense: "defense",
} as const;

export interface PartStats {
  armor: number;
  power: number;
  speed: number;
  energy: number;
}

export interface RobotPart {
  id: string;
  name: string;
  category: RobotPartCategory;
  stats: PartStats;
  color: string;
  description: string;
  /** @nullable */
  specialAbility?: string | null;
}

export interface Robot {
  id: number;
  name: string;
  playerName: string;
  bodyPartId: string;
  attackPartId: string;
  defensePartId: string;
  totalStats: PartStats;
  /** @nullable */
  specialAbility?: string | null;
  createdAt: string;
}

export interface RobotInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  playerName: string;
  bodyPartId: string;
  attackPartId: string;
  defensePartId: string;
}

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus];

export const RoomStatus = {
  waiting: "waiting",
  fighting: "fighting",
  finished: "finished",
} as const;

export interface Room {
  id: string;
  name: string;
  status: RoomStatus;
  playerCount: number;
  maxPlayers: number;
  createdAt: string;
  /** @nullable */
  hostName?: string | null;
}

export interface RoomInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  hostName: string;
}

export interface LeaderboardEntry {
  rank: number;
  playerName: string;
  wins: number;
  losses: number;
  winRate: number;
  totalBattles: number;
  /** @nullable */
  favoriteRobot?: string | null;
}