# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Robo Arena API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /robots:
    get:
      operationId: listRobots
      tags: [robots]
      summary: List all saved robot builds
      responses:
        "200":
          description: List of robots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Robot"
    post:
      operationId: createRobot
      tags: [robots]
      summary: Save a robot build
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RobotInput"
      responses:
        "201":
          description: Robot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Robot"
  /robots/{id}:
    get:
      operationId: getRobot
      tags: [robots]
      summary: Get a specific robot build
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Robot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Robot"
        "404":
          description: Not found
...
components:
  schemas:
...
    PartStats:
      type: object
      required: [armor, power, speed, energy]
      properties:
        armor:
          type: integer
        power:
          type: integer
        speed:
          type: integer
        energy:
          type: integer
    Robot:
      type: object
      required: [id, name, playerName, bodyPartId, attackPartId, defensePartId, totalStats, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        playerName:
          type: string
        bodyPartId:
          type: string
        attackPartId:
          type: string
        defensePartId:
          type: string
        totalStats:
          $ref: "#/components/schemas/PartStats"
        specialAbility:
          type: ["string", "null"]
        createdAt:
          type: string
    RobotInput:
      type: object
      required: [name, playerName, bodyPartId, attackPartId, defensePartId]
      properties:
        name:
          type: string
          minLength: 1
        playerName:
          type: string
          minLength: 1
        bodyPartId:
          type: string
        attackPartId:
          type: string
        defensePartId:
          type: string
...