# lib/api-zod/src/generated/types/partStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robo Arena API
 * OpenAPI spec version: 0.1.0
 */

export interface PartStats {
  armor: number;
  power: number;
  speed: number;
  energy: number;
}