# lib/api-zod/src/generated/types/robot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robo Arena API
 * OpenAPI spec version: 0.1.0
 */
import type { PartStats } from "./partStats";

export interface Robot {
  id: number;
  name: string;
  playerName: string;
  bodyPartId: string;
  attackPartId: string;
  defensePartId: string;
  totalStats: PartStats;
  /** @nullable */
  specialAbility?: string | null;
  createdAt: string;
}