# lib/api-zod/src/generated/types/robotInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robo Arena API
 * OpenAPI spec version: 0.1.0
 */

export interface RobotInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  playerName: string;
  bodyPartId: string;
  attackPartId: string;
  defensePartId: string;
}